@@ -5,22 +5,45 @@
   "main": "webpack.config.js",
   "dependencies": {
     "@types/createjs": "0.0.29",
-    "@types/js-cookie": "^2.1.0",
+    "@types/jquery": "^3.3.12",
+    "@types/js-cookie": "^2.2.0",
+    "@types/js-yaml": "^3.11.2",
     "@types/node-uuid": "0.0.28",
-    "@types/socket.io": "^1.4.31",
+    "@types/socket.io": "^1.4.38",
     "@types/socket.io-client": "^1.4.32",
-    "express": "^4.16.2",
+    "express": "^4.16.3",
+    "jquery": "^3.3.1",
     "js-cookie": "^2.2.0",
+    "js-yaml": "^3.12.0",
     "node-uuid": "^1.4.8",
-    "router": "^1.3.2",
-    "socket.io": "^2.0.4",
-    "ts-loader": "^3.2.0",
-    "webpack": "^3.10.0"
+    "npm": "^6.4.1",
+    "router": "^1.3.3",
+    "socket.io": "^2.1.1",
+    "ts-loader": "^5.2.1",
+    "typescript": "^3.1.1"
   },
-  "devDependencies": {},
+  "devDependencies": {
+    "@types/jasmine": "^2.8.9",
+    "gulp": "^4.0.0",
+    "gulp-imagemin": "^4.1.0",
+    "gulp-typescript": "^5.0.0-alpha.3",
+    "gulp-watch": "^5.0.1",
+    "gulp.spritesmith": "^6.9.0",
+    "hard-source-webpack-plugin": "^0.12.0",
+    "imagemin-pngquant": "^6.0.0",
+    "jasmine": "^3.2.0",
+    "require-dir": "^1.1.0",
+    "run-sequence": "^2.2.1",
+    "webpack": "^4.20.2",
+    "webpack-cli": "^3.1.2",
+    "webpack-stream": "^5.1.1"
+  },
   "scripts": {
     "start": "node dist/Server/main.js",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "npx jasmine",
+    "postinstall": "gulp compress && gulp sprite && gulp sprite_compress && gulp webpack_pro && gulp build",
+    "build": "webpack && tsc -p Server && npm start",
+    "build_only": "gulp test_build"
   },
   "repository": {
     "type": "git",
