@@ -1,2 +1,7 @@
-export const canvasWidth = 960;
-export const canvasHeight = 876;
+
+export module global {
+    export const canvasWidth = 960;
+    export const canvasHeight = 876;
+    export const cardIconSize = 30 * 2;
+    export const costCountMax = 5;
+}
