@@ -2,10 +2,8 @@
 import * as http from 'http';
 import { createRouter } from "./routing";
 import * as io from "socket.io";
-import { LoginControler } from "../Server/loginControler";
-import { LoginSocket } from "../Server/loginSocket";
-import { BoardControler } from "../Server/boardControler";
-import { BoardSocket } from "../Server/boardSocket";
+import { createControler } from "./createControler";
+
 
 //サーバーの作成
 let server = http.createServer();
@@ -14,11 +12,6 @@ let server = http.createServer();
 server.on("request", createRouter());
 //メインソケットを作成
 const mainSocket = io(server);
-//メインソケットからログインソケットを作成
-const boardControler = new BoardControler;
-
-const loginSocket = new LoginSocket(mainSocket,boardControler);
-const boardSocket = new BoardSocket(mainSocket,boardControler);
-
+createControler(mainSocket);
 //サーバーポート設定
-server.listen(3000);
+server.listen(process.env.PORT || 3000);
