@@ -14,7 +14,7 @@ export class PlayerDataList {
     }
 
     public getPlayerData(uuid: string) {
-        return this.playerDataList.find(x => {
+        return <PlayerData | undefined>this.playerDataList.find(x => {
             if (x == null) return false;
             return x.getUuid() == uuid;
         });
