@@ -1,11 +1,6 @@
-import { RoomEvents } from "../Server/roomEvents";
-import { PlayerData } from "../Server/playerData";
-import { PlayerDataList } from "../Server/playerDataList";
-import { PlayFlagDataForClient } from "../Share/playFlagDataForClient";
-import { PlayerDataForClient } from "../Share/playerDataForClient";
+import { PlayerData } from "./playerData";
+import { PlayerDataList } from "./playerDataList";
 import { PasswordInfo } from "./passwordInfo";
-import { PassThrough } from "stream";
-import { RoomDataForClient } from "../Share/roomDataForClient";
 
 export class RoomData {
     private roomId: number;
@@ -13,62 +8,37 @@ export class RoomData {
     private playerDataList: PlayerDataList;
     private playFlag: boolean;
     private passwordInfo: PasswordInfo;
-    private roomEvents: RoomEvents;
 
-    constructor(roomId: number, roomName: string, passwordInfo: PasswordInfo, roomEvents: RoomEvents) {
+    constructor(roomId: number, roomName: string, passwordInfo: PasswordInfo, /*roomEvents: RoomEvents*/) {
         this.roomId = roomId;
         this.roomName = roomName;
         this.passwordInfo = passwordInfo;
         this.playFlag = false;
-        this.roomEvents = roomEvents;
 
         this.playerDataList = new PlayerDataList;
-
-        let roomData: RoomDataForClient = {
-            roomName: this.roomName,
-            roomId: this.roomId,
-            playFlag: this.playFlag,
-            playerList: this.playerDataList.getPlayerNameList(),
-            passwordFlag: this.passwordInfo.isNeedPassword()
-        };
-        this.roomEvents.addRoom(roomData);
     }
 
     getRoomId() { return this.roomId; }
 
     getRoomName() { return this.roomName; }
 
-    getPlayFlag() { return this.playFlag; }
+    get PlayFlag() { return this.playFlag; }
 
-    setPlayFlag(playFlag: boolean) {
+    set PlayFlag(playFlag: boolean) {
         this.playFlag = playFlag;
-        let playFlagDataForClient: PlayFlagDataForClient
-            = { playFlag: playFlag, roomId: this.roomId };
-        this.roomEvents.updatePlayFlag(playFlagDataForClient);
     }
 
-    getPlayerData(uuid: string) { this.playerDataList.getPlayerData(uuid); }
+    getPlayerData(uuid: string) {
+        return this.playerDataList.getPlayerData(uuid);
+    }
 
     deleteMember(uuid: string) {
         this.playerDataList.deleteMember(uuid);
-        const playerDataForClient: PlayerDataForClient =
-            {
-                roomId: this.roomId,
-                playerId: this.playerDataList.getPlayerId(uuid),
-                playerName: ""
-            };
-        this.roomEvents.deleteMember(playerDataForClient, uuid);
     }
 
     addMember(playerData: PlayerData) {
         this.playerDataList.addMember(playerData);
-        const playerDataForClient: PlayerDataForClient =
-            {
-                roomId: this.roomId,
-                playerId: this.playerDataList.getPlayerId(playerData.getUuid()),
-                playerName: playerData.getName()
-            };
-        this.roomEvents.addMember(playerDataForClient);
+        return this.playerDataList.getPlayerId(playerData.getUuid());
     }
 
     getPasswordInfo() { return this.passwordInfo; }
@@ -76,7 +46,6 @@ export class RoomData {
     deleteRoom() {
         this.playerDataList.getPlayerUuidList()
             .forEach(x => this.deleteMember(x));
-        this.roomEvents.deleteRoom(this.roomId);
     }
 
     getPlayerCount() { return this.playerDataList.getPlayerCount(); }
