@@ -1,16 +1,16 @@
-export class RoomIdGenerator{
-    private unUsedId:number[];
-    constructor(){
+export class RoomIdGenerator {
+    private unUsedId: number[];
+    constructor() {
         this.unUsedId = new Array(100);
-        for(let i=0;i<this.unUsedId.length;i++){
+        for (let i = 0; i < this.unUsedId.length; i++) {
             this.unUsedId[i] = (i + 1);
         }
     }
-    getRoomId(){
+    getRoomId() {
         return this.unUsedId.length > 0 ? this.unUsedId.pop()! : null
     }
-    releaseRoomId(roomId:number){
-        if(this.unUsedId.find(x => x == roomId) == undefined){
+    releaseRoomId(roomId: number) {
+        if (this.unUsedId.find(x => x == roomId) == undefined) {
             this.unUsedId.push(roomId);
         }
     }
