@@ -1,6 +1,10 @@
 
 console.log("すごーい！君はwebpackできるフレンズなんだね！");
+const HardSourceWebpackPlugin = require('hard-source-webpack-plugin');
 module.exports = {
+
+    mode: 'development',
+
     entry: {
         "Client/board": "./Client/board.ts",
         "Client/login": "./Client/login.ts",
@@ -15,7 +19,10 @@ module.exports = {
     module: {
         rules: [
             // all files with a `.ts` or `.tsx` extension will be handled by `ts-loader`
-            { test: /\.tsx?$/, loader: 'ts-loader' }
+            { test: /\.tsx?$/, use: 'ts-loader' }
         ]
-    }
+    },
+    plugins: [
+        new HardSourceWebpackPlugin()
+    ]
 }
